% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_ensemble.R
\name{dimension_ensemble}
\alias{dimension_ensemble}
\title{model ensemble}
\usage{
dimension_ensemble(yraw, alpha, gg, kk, ll, pp, cores_number, hh,
  prior_constant_variance, density_size, sub, dimensions)
}
\arguments{
\item{yraw}{data}

\item{alpha}{forgetting factor}

\item{gg}{state variance}

\item{kk}{observation variance}

\item{ll}{prior parameter}

\item{pp}{lags}

\item{cores_number}{number of parallel cores}

\item{hh}{forecast horizon}

\item{prior_constant_variance}{expecttaion prior choice}

\item{density_size}{density_size}

\item{sub}{sub models for dms}

\item{dimensions}{variable to forecast}
}
\value{
forecasts and probability of the models
}
\description{
forecast using dms
}

