% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ensemble.R
\name{model_ensemble}
\alias{model_ensemble}
\title{main function}
\usage{
model_ensemble(gg = 0.1, kk = 1, ll = 1, cores_number = 1)
}
\arguments{
\item{gg}{prior parameter}

\item{kk}{observe variance parameter}

\item{ll}{state variance parameter}

\item{cores_number}{define the number of cores}
}
\value{
the return
}
\description{
this is the description
}
\examples{
model_ensemble()
}

