% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock_watson_forecast_errors}
\alias{stock_watson_forecast_errors}
\title{stock and watson forecast errors}
\format{An object of class \code{list} of length 39.}
\usage{
stock_watson_forecast_errors
}
\description{
stock and watson forecast errors
}
\keyword{datasets}

