% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock_watson_forecast_errors_benchmark_ar}
\alias{stock_watson_forecast_errors_benchmark_ar}
\title{stock and watson forecast errors benchmark ar}
\format{An object of class \code{list} of length 39.}
\usage{
stock_watson_forecast_errors_benchmark_ar
}
\description{
stock and watson forecast errors benchmark ar
}
\keyword{datasets}

