% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock_watson_raw}
\alias{stock_watson_raw}
\title{stock and watson raw data}
\format{An object of class \code{matrix} with 164 rows and 525 columns.}
\usage{
stock_watson_raw
}
\description{
stock and watson raw data
}
\keyword{datasets}

