% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock_watson_transformed}
\alias{stock_watson_transformed}
\title{stock and watson transformed data}
\format{An object of class \code{matrix} with 164 rows and 525 columns.}
\usage{
stock_watson_transformed
}
\description{
stock and watson transformed data
}
\keyword{datasets}

